requirements.txt
streamlit>=1.35.0
PyMuPDF>=1.24.0
langchain>=0.2.0
langchain-community>=0.2.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
google-genai>=1.5.0
python-dotenv>=1.0.0
numpy>=1.26.0